IOT-1375 Add OC_DEPRECATED macros
This add macros that enable marking deprecated for the compiler
This will add compiler specific annotations that will
cause compiler warnings when a deprecated function is used.

Adding the date the function was deprecated, was inspired by
the AllJoyn project. Adding the date a function or class was
deprecated can help when deciding to remove deprecated code.

The doxygen config files were modified to expand the macros
so doxygen properly documents the deprecated function not the
OC_DEPRECATED macro itself.

The macro was then used to mark the base64 implementation
deprecated. Since the base64 implementation is currently placed
in the public header files we should mark it as deprecated giving
other users time to move their code to the mbedtls
implementation.

Change-Id: I6e13e34caf84311e555d19aeb393e922d09d4b59
Signed-off-by: George Nash <george.nash@intel.com>