IOT-582 Android build generate universal APK
Two major modifications:
1) The aar file no longer adds the TARGET_ARCH to the
name of the file. The aar file contains all of the
architectures that were built upto that point. This
has been the case for quite a while so the architecture
in the name only indicated the build that created the
aar file not the contents of the aar file.

2) When APK files are built it will try to build all
possible architectures as well as the universal APK
file.

Bug: https://jira.iotivity.org/browse/IOT-582
Change-Id: I11f5db9f4cb8c634afe4417b280e5feec6c3d7e9
Signed-off-by: George Nash <george.nash@intel.com>