[IOT-2906][ES] Remove "href" from child response.
For requests which come directly to child resources
(/WiFiConfResURI, /CoapCloudConfResURI and /DevConfResURI),
this patch will remove the "href" property from the response
as spec doesn't mandate the addition of "href" in the response.

And also, CTT expects the response from direct request to any child
to be same as the response which comes for the same child
through collection.

For example:
Properties in "rep" for '/WiFiConfResURI' should be same as
the properties in the direct response for GET from '/WiFiConfResURI' resource.

Currently "href" is not included in "rep" for child resources.
But it is included in the direct response for GET/POST.
When CTT compares both responses, it finds them different due of "href".

This patch resolves the issue and still conforms to spec.

Change-Id: I017bed18d3e0ad568b662a03d9bd80134538a8c0
Signed-off-by: Senthil Kumar G S <senthil.gs@samsung.com>