Detailed PIN logging during RDP OTM
At times duplicate Random PIN callbacks are displayed at IUTSimulator
app console for a single PIN prompt at CTT. It creates a confusion for
test operator and increases the chances for entering wrong PIN. 

I modified the sample app to display a local sequence number and time
stampt whenever PIN callback is received. Also PIN CODEs are stored
in a separate log file so it makes it easy to track PIN when IUT is 
built in debug mode.

Change-Id: If128e3dfd77bcd4fea1c5b002fc0bea2301aefda
Signed-off-by: Abhishek Pandey <abhi.siso@samsung.com>