Remove arduino support 1/3
Arduino is now supported only in iotivity-constrained,
but all the support in the iotivity build remains.
Remove in three phases (just to make shorter commits
to review).  This patch removes from the top level
(root of the tree and build_common).

Assumption: the logic to avoid multi-transport and the
logic to build without dtls support were both written
in a general way, but the list of targets in both cases
contained just 'arduino', so this logic was dropped.

Change-Id: Ib120ab7d82ff91b5f744c638d31f1b430a50ccd7
Signed-off-by: Mats Wichmann <mats@linux.com>