Remove arduino support 2/3
Arduino is now supported only in iotivity-constrained,
but all the support in the iotivity build remains.
Remove in three phases (just to make shorter commits
to review).  Part 2 removes arduino from all the other
build scripts which referenced it, and from auto_build.py
and the vagrant setup.

Change also removes the remaining arduino directories, as
follows:

    extlibs/arduino
    resource/c_common/ocatomic/src/arduino
    resource/c_common/ocrandom/test/arduino
    resource/csdk/connectivity/build/arduino
    resource/csdk/connectivity/lib/arduino
    resource/csdk/connectivity/src/bt_le_adapter/arduino
    resource/csdk/connectivity/src/ip_adapter/arduino
    resource/csdk/connectivity/src/tcp_adapter/arduino
    resource/csdk/logger/test/arduino
    resource/csdk/stack/samples/arduino
    resource/csdk/stack/test/arduino
    service/easy-setup/sampleapp/enrollee/arduino
    tools/arduino

which means some code files are removed as well.

Bug: https://jira.iotivity.org/browse/IOT-2927
Change-Id: Idd5001a810ff4f67022f60bcb9d3252f7de7ccc2
Signed-off-by: Mats Wichmann <mats@linux.com>