Squash gcc7 warnings (-> fails)
base64 stuff in iotivity is deprecated, but while it is
only deprecated, not removed, the unit tests should
continue to exist and run.  Thus, the one test source file
should not fail the build due to emitting warnings over
the use of deprecated features.  The pragma works only
for gcc, so protect it from other compilers.

Change-Id: Id5d375c7d0bb601d87134f701d0da9e363c9fd76
Signed-off-by: Mats Wichmann <mats@linux.com>