Align license file
Minor:
The current LICENSE.md file is modified from the Apache original.
The modification is mostly incorrect (the change made is
intended for other source files, as described in the Appendix
in the original, not for the license file itself).  Restored the
pristine original.

Main problem addressed: github mirror misidentifies the license
of the project, since the first few lines do not match the
known text of the Apache 2.0 license.

Change-Id: Ia994002384bf93fa999ac634bc6d4c711b5c9d60
Signed-off-by: Mats Wichmann <mats@linux.com>