Sample implementation of ocf_light
This sample borrows a lot of code from the server
sample generated from the DeviceBuilder project
https://github.com/openconnectivityfoundation/DeviceBuilder

Eventhough this borrows a lot of code from the
DeviceBuilder project it has been modified to make
it more useful as an example separating the code into
smaller classes and adding helper classes.

This code has been run against the OCF CTT. See the
README.md file for more information about the CTT
results.

Change-Id: If0ae5e3527517bb722648a614fbd96f13ad01144
Signed-off-by: George Nash <george.nash@intel.com>