static analysis: buffer size violations
fix reported "might leave destination string unterminated" when
using strncpy and the size argument is the size of the desitnation
buffer. There are several ways to fix, but the IoTivity API OCStrncpy
adjusts the size before calling the underlying fuction and so fixes
the problem. This is the coding standard recommendation anyway.

for instances in examples, oic_string.h is not in the public API,
so just fix up the counts.

bridging/common/messageHandler.cpp had a third strncpy which was not
flagged - but it was using a constant that did not match the size
in the destination - MPMResourceList.href is size MPM_MAX_URI_LEN,
so this was adjusted.

service/notification/src/consumer/NSConsumerCommon.c had this construct:
   sizeof(char) * NS_DEVICE_ID_LENGTH
in several places (one of which was one of these strncpy calls that
was changed). the instances were shortened NS_DEVICE_ID_LENGTH for
readability. This is not fixing any reported problem and can be
dropped if it bothers reviewers.

Change-Id: I8f22f7dd704849477dad0dd1f16cd9276ebf1d04
Signed-off-by: Mats Wichmann <mats@linux.com>