coap-http-proxy: Sync JSON and CBOR files
File is updated after running test.

To avoid local changes on build,
let's commit the updated .dat files (using json2cbor),

Reminding, It's developer dutty to commit those files on each changes
on json source or any change that produce a different result.
This way we're not leaving inconsistant binary files in tree.

JSON file was not modified because it is already in UNIX format

Note: those files may need to be updated to support current ACL
(and commited by security team, check status of json and dat files)

Bug: https://jira.iotivity.org/browse/IOT-2310
Change-Id: I07a7482eab8eef04c26604e0113020d5c62603a1
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>