Annotate deprecated functions with macro
If a function was already marked as deprecated using
the @derecated doxygen annotation the OC_DEPRECATED_MSG
macro was added to generate compiler warnings when a
deprecated function was used.

The date the function was deprecated was obtained by
using git blame to find when the @deprecated tag
was added to the function documentation. The date the
tag was added was used in the macro.

Three instances of the function registerDeviceInfo
are marked as deprecated but the macro was not added
due to inheritance and mock objects implementing
their own version of the registerDeviceInfo. Caused
difficulty adding the macro to them without breaking
the build.

An attempt was made to clean up any remaining usages
of deprecated functions that were not part of unit tests.
It is possible that one or more instances of the
'registerDeviceInfo' function was missed.

Java is still using deprecated functions:
IOT-3011 and IOT-3012 to address the fact that Java is
still using deprecated functions.

Change-Id: I39b1af690b99d52f2e0d026e4c27b908e2f6c052
Signed-off-by: George Nash <george.nash@intel.com>