Update sampleserver_justworks .json and .dat
This is the first of the applications to recieve an update to
the /acl2 .json to utilize the new "resources" wildcards.

This change (adding the two "+" and "-" entries in /acl2) should probably
be propagated to most sample applications, to give a consistent
example for vendors assessing IoTivity samples, but it is
not required for correction function.  The exception is an app
that is using the "*" wildcard to grant access to DCRs (e.g. the
SVRs) which will no longer work (but shouldn't have been done
anyway!)

Change-Id: I66e06432dead4ec90c17091352d2c19cab753bcf
Signed-off-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>