IOT-3045 Fix return type of getResourceHandleAtUri
OcPlatfrom.getResourceAtUri should return null
when a resource is not found at the specified
URI. The class JniOcResourceHandle returned
a non-null value even when a null resourceHandle
was passed into the function. The assumption
when the code was originally created was that
it would return null. This assumption was incorrect.

This was tested by modifying simpleserver sample
Modifications not included as part of this commit.

Bug: https://jira.iotivity.org/browse/IOT-3045
Change-Id: I3e923dca3415e9e0ec473b1dd4ff4619538f325b
Signed-off-by: George Nash <george.nash@intel.com>