Fix two more problems building with Python3
The boost script used string.replace() which is no longer
supported; problem did not turn up until trying a build on Windows

The security - provisioning - unittest script had a Windows
stanza for killing running server processes before launching a
new one. It printed the process id it was going to kill in a way
that caused an AttributError exception - this was a conversion
error from when print statement was changed to print function,
again the code path had not been hit. At the same time a second
exception ocurred because when the first exception took place the
script attempted to us the message attribute of the exception for
printing, but not all exceptions have such an attribute. Change
to just print the string representation of the exception, which
is more reliable.

Additionally, in the libcoap script, the term 'string' is used
which while not broken since the script never imports the string
module, is unnecessarily ambiguous. Use the Python recommended
disambiguation of adding a trailing underscore on an identifier
which might match one used by Python.

Change-Id: I635796a6442caacb6988e21dbe46938f860e333f
Signed-off-by: Mats Wichmann <mats@linux.com>