fix compilation failure on VS2017
In the latest build of Visual Studio 2017 will
produce a build waring when encountering any code
in the str::tr1 namespace. The warning is produced
when headers from the external googletest framework
are included in the build.

Adding the _SILENCE_TR1_NAMESPACE_DEPRECATION_WARNING
macro will cause the build to ignore the build warning.

The googletest project is aware of the issue. See the
actual commit for a link to the ticket for the google
test project.

Change-Id: I9be616498df95b8dd9aa165717bd178c53f3221b
Signed-off-by: George Nash <george.nash@intel.com>