webOS: Fix IoTivity Sampler bug
:Release Notes:
Fix IoTivity Sampler bug

:Detailed Notes:
Application name of the IoTivity Sampler application is changed.
And the IoTivity Sampler application can now be closed using
'x' button. Also, there was a bug that observe request does
not work due to typoe of length function. The typo is fixed in
this commit. Refer to IOT-3181 for more information about the
bug.

:Testing Performed:
Verified that name of the application is changed.
Verified that IoTivity Sampler is closed using 'x' button
Verified that observer request works as expected

:Issues Addressed:
[IOT-3181] webOS: Fix IoTivity Sampler bug

Change-Id: If38692b52cde62db6dd9515192b56e0598532560
Signed-off-by: sungmok.shin <sungmok.shin@lge.com>