Update upstream-libcoap to IoTivity-1.4 tag
New coap code is needed so update the tag to look for. Fixes the
failed Windows CI builds in https://gerrit.iotivity.org/gerrit/#/c/25491/

Other changes:
* The SConscript adds the WITH_WS definition to the build
* Quiet the windows warning on size mismatches (W4267)
- Note these two indicate problems introduced by iotivity patches to
the dthaler libcoap fork, and the problems themselves should be fixed
(WITH_WS is correct, but the logic in the case statement is not, so the
absence of the flag caused build failures which are now covered up -
this one is IOT-3103)

* libcoap script checked tag in last line of commit log, that's fragile,
so changed to use git log instead.

* The standard check for the alternate way the code may be populated
is added.  This is used for Tizen builds.

* A little restructuring for readability (message strings defined away
from the code so we can read the code), and the "git checkout" is now
actually error-checked.

Bug: https://jira.iotivity.org/browse/IOT-3044
Change-Id: I331ef9ed19d128ff975d150f6c0058eff64b59ff
Signed-off-by: Mats Wichmann <mats@linux.com>