[IOT-2234] Fix libcoap configure for Windows UWP
The libcoap sconscript has a special case for Windows UWP because it
did not configure right. The problem is the list of libraries set by the
Windows build_common script is cleared before running the configure test,
and that is not the right answer for UWP, which needs one particular
library in order to access the standard Windows library functions.
Rather than building the exact knowledge into this script, the required
list is saved in the Windows build script, and retrieved here for use
for the UWP platform.

Bugs: https://jira.iotivity.org/browse/IOT-2234
Change-Id: I70016d81c5d51c7454b1794280e12f5574a5fd41
Signed-off-by: Mats Wichmann <mats@linux.com>