valgrind default switched to off.
A construction variable, VALGRIND_CHECKS, is added to the environment if
the target supports valgrind (currently that list is is linux and darwin).
It defaults to True.  If TEST mode is on, then the various unit test
locations will call the run_tests() tool which builds a command line
to run the test. If two conditions are met - (1) a pathname indicating
where to store the valgrind output is passed as the second argument
and (2) VALGRIND_CHECKS is set in the environment and evaluates True,
then that command line will be written to run the test under valgrind
control. In practical terms, This means the (linux) unit_tests builder
will run tests under valgrind. This happens on every patch submitted
to Gerrit, as part of the Jenkins CI setup.

At the moment, valgrind reports are not actively pursued by developers,
and are proving unstable (that is the terminology used by the Jenkins
valgrind report plugin itself, indicating that results differ from run
to run). Many patches have been stalled by the valgrind report plugin
marking the unit_tests build a failure, even if the patch had no
impact at all on the run.

Until this situation improves, flip the VALGRIND_CHECKS default to
False. Any developer can set up their own valgrind run at any time
by passing TEST=1 VALGRIND_CHECKS=1 to scons.

The run_test docstring is reworded.

Change-Id: I2519ab8771e2a9e299853a0e1574a10568db68a2
Signed-off-by: Mats Wichmann <mats@linux.com>