examples/OCFSecure: Add ARTIK board
This patch add the ARTIK board to the list of supported devices.
The blue LED on board (GPIO38) is used for switch resource in the server
program. And the GPIO is controlled by "mraa" library which already used
in the server program.

The "mraa" library is not support the ARTIK board. So to use the raw GPIO
pin number, "raw=true" option is added.

Change-Id: I2004bc473c64189f3cc1010ae76c4cfe757b33f5
Signed-off-by: Inho Oh <webispy@gmail.com>