Separate non-core tizen builds for Jenkins
The continuous integration system for IoTivity builds the targets
tizen_secured and tizen_unsecured as defined in auto_build.py.  These two
not only do a full build of tizen, but also several extra builds of
samples as defined in the build_tizen function.  This change separates the
extra builds into separate targets so a developer can still easily
call them, but they will not be part of the required CI build on
every change submitted to gerrit.

Change-Id: I1c8846e74afb2eb18532baa64543e320ef045ba9
Signed-off-by: Mats Wichmann <mats@linux.com>