Fixed Android OcPlatform.Shutdown() does not work.
Android OcPlatform.Shutdown() API does not work properly, because in
the case of Android, start() is called twice: once by the Android
Configure API, which is normal and needed, and once by the constructor
for OcPlatform_impl()!
This last one is due to the fact that useLegacyCleanup is still set
to true in the PlatformConfig used by Android, which is one of the
2 non-deprecated ones.
Comments say that the flag is set to false by default, unless a
legacy constructor is used. But this is obviously not a legacy
constructor, and the flag was erroneously set to true...
With this fix, Android OcPlatform.Shutdown() works as expected.

Change-Id: I3db4c37ff17c9f16ea927093c8aa2806e1cabb90
Signed-off-by: Herve Jourdain <herve.jourdain@beechwoods.com>