[Easy Setup] Reject UPDATE on not allowed property
[IOT-3231] https://jira.iotivity.org/browse/IOT-3231
Easy Setup Resource does not reject an UPDATE request to add a new Link

Currently only criteria to reject POST on EasySetup Resource
(/EasySetupResURI) is to check if properties are read-only. However,
the POST request may contain additional properties "href", "rep"
representing a link to be added in collection or any other unsupported
properties. Even though these properties are not updated, success
response is sent to client which is not appropriate and is flagged as
CTT failure.

This patch changes the validation logic on /EasySetupResURI for POST
on baseline interface to allow the POST only if payload contains
the writable property "cn". Update to any other property is rejected.

Change-Id: I33f6e7eda33481fa2137341f93b607a837ff5a87
Signed-off-by: Abhishek Pandey <abhi.siso@samsung.com>