[Easy Setup] Check required properties for UPDATE
[IOT-3252]
https://jira.iotivity.org/browse/IOT-3252

Per the WiFiConf Schema, for an UPDATE request:
 "required":["tnn", "wat", "wet"]

This means that all three of these Properties must be in every
UPDATE request. If they are not there, the Device is supposed
to reject the request with a 4.xx error. However, IoTivity
responds with a 2.04 Changed response instead currently.

This patch fixes above issue by checking the POST request
for required properties (tnn, wat, wet). If any of these
are not included in POST request, it will be rejected.

Change-Id: I00f02c93cd273a02706c85c5d236584d45d6ca77
Signed-off-by: Abhishek Pandey <abhi.siso@samsung.com>