Show how to make examples non-default
A scheme is introduced which allows separating targets indended to
be built by default from ones not intended to be built by default.
One example is converted to show how it works.

Since this makes use of added methods (extending scons' own API),
took the opportunity to properly document these, and added a
documentation file in a new doc subdirectory to make it easier
for developers to find out about these methods.

Also added a debug scheme, which makes it more clear where to
put tracing statements in between the two passes of scons -
the script is invoked at the very end of the scan pass, before
the build pass begins but when everything should be defined.

Change-Id: I0a636fa34c18ace23720f22bfdd8511afa611ddb
Signed-off-by: Mats Wichmann <mats@linux.com>